+++ python/world.cpp
@@ -24,11 +24,7 @@ void init_world(py::module_ &m) {
         .def("cleanup", &World::cleanup)
         .def("cleanupDEM", &World::cleanupDEM)
         .def("cleanupDOM", &World::cleanupDOM)
-
-        .def("getRayResolution", &World::getRayResolution)
-        .def("setRayResolution", &World::setRayResolution)
         
         .def("computeRayResolution", &World::computeRayResolution);
-
 
 }